Rename GroupByKeyEvaluator to GroupByKeyOnlyEvaluator, etc
